+++ fsutil/fsutil.go
@@ -4,16 +4,16 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
+	"time"
 
 	"github.com/mitchellh/go-homedir"
 )
 
-// DirWritable checks the the directory is writable.  If the directory does
+// DirWritable checks if a directory is writable. If the directory does
 // not exist it is created with writable permission.
 func DirWritable(dir string) error {
 	if dir == "" {
-		return errors.New("cannot check empty directory")
+		return errors.New("directory not specified")
 	}
 
 	var err error
@@ -25,16 +25,15 @@ func DirWritable(dir string) error {
 	_, err = os.Stat(dir)
 	if err == nil {
 		// dir exists, make sure we can write to it
-		testfile := filepath.Join(dir, "test")
-		fi, err := os.Create(testfile)
+		file, err := os.CreateTemp(dir, "test")
 		if err != nil {
 			if os.IsPermission(err) {
 				return fmt.Errorf("%s is not writeable by the current user", dir)
 			}
-			return fmt.Errorf("unexpected error while checking directory writeablility: %w", err)
+			return fmt.Errorf("error while checking directory writeablility: %w", err)
 		}
-		fi.Close()
-		return os.Remove(testfile)
+		file.Close()
+		return os.Remove(file.Name())
 	}
 
 	if os.IsNotExist(err) {
@@ -47,6 +46,19 @@ func DirWritable(dir string) error {
 	}
 
 	return err
+}
+
+// FileChanged returns the modification time of a file and true if different
+// from the given time.
+func FileChanged(filePath string, modTime time.Time) (time.Time, bool, error) {
+	fi, err := os.Stat(filePath)
+	if err != nil {
+		return modTime, false, fmt.Errorf("cannot stat file %s: %w", filePath, err)
+	}
+	if fi.ModTime() != modTime {
+		return fi.ModTime(), true, nil
+	}
+	return modTime, false, nil
 }
 
 // FileExists return true if the file exists
